+++ src/repositories/attachment_repository.go
@@ -13,7 +13,7 @@ var (
 
 type AttachmentRepository interface {
 	Create(ctx context.Context, attachment *entities.Attachment) error
-	GetByID(ctx context.Context, id int64) (*entities.Attachment, error)
-	GetByMessageID(ctx context.Context, messageID int64) ([]*entities.Attachment, error)
-	DeleteByMessageID(ctx context.Context, messageID int64) error
+	GetByID(ctx context.Context, id uint) (*entities.Attachment, error)
+	GetByMessageID(ctx context.Context, messageID uint) ([]*entities.Attachment, error)
+	DeleteByMessageID(ctx context.Context, messageID uint) error
 }
