+++ src/repositories/recipient_repository.go
@@ -13,6 +13,6 @@ var (
 
 type RecipientRepository interface {
 	Create(ctx context.Context, recipient *entities.Recipient) error
-	GetByMessageID(ctx context.Context, messageID int64) ([]*entities.Recipient, error)
-	DeleteByMessageID(ctx context.Context, messageID int64) error
+	GetByMessageID(ctx context.Context, messageID uint) ([]*entities.Recipient, error)
+	DeleteByMessageID(ctx context.Context, messageID uint) error
 }
